TIC TAC TOE
Three-dimensional tictactoe is played on a 4 x 4 x 4 board. As in
ordinary tictactoe, players X and O mark squares alternately. The winner
is the first to align four of his/her own marks in any direction,
parallel to one side of the board or along a diagonal. There are 76 ways
of wining. Write a program to play the game against a human opponent.