Tournament A
You are to organize a tournament involving n competitors. Each
competitor must play exactly once against every other competitor.
Moreover, each competitor must play exactly one match every day, with
the possible exception of a single day when he/she does not play at all.
(a) If n is a power of 2, write a program to construct a timetable
allowing the tournament to be finished in n-1 days.