YouTube Video and Playlist Player
This is a YouTube Video and Playlist player that can play any video or
playlist based on the ID you provide. There are two java files where
Id's have to be added and those are youtubeactivity.java and
submenu.java. Also to run this app use your own API key. It is build on
Android Studio.