Using Ajax and JSON data format
A server side program using Java servlets that is invoked using AJAX
making use of JSON to communicate some user action on a web page and
make use of DOM functions and JavaScript to incorporate the results
returned by AJAX into the existing page. A full array of CSS
functionality – demonstrating cascades and use of classes is
implemented.